Adjusts code review

+++ application/src/main/java/com/mercadolibre/planning/model/me/gateways/outboundwave/OutboundWaveGateway.java
@@ -4,7 +4,7 @@ import java.time.ZonedDateTime;
 
 public interface OutboundWaveGateway {
 
-    long getUnitsCount(final String warehouseId,
+    int getUnitsCount(final String warehouseId,
                        final ZonedDateTime dateFrom,
                        final ZonedDateTime dateTo,
                        final String unitGroupType);
